--- src/main/java/com/woncheonpeople/shoppingmall/domain/Member.java
@@ -1,22 +1,0 @@
-package com.woncheonpeople.shoppingmall.domain;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Entity(name = "members")
-@Getter
-@Setter
-public class Member {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String name;
-
-    public Member(String name) {
-        this.name = name;
-    }
-}
