--- src/main/java/com/woncheonpeople/shoppingmall/domain/Product.java
@@ -1,23 +1,0 @@
-package com.woncheonpeople.shoppingmall.domain;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Entity(name = "products")
-@RequiredArgsConstructor
-@Getter
-@Setter
-public class Product {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String name;
-
-    public Product(String name) {
-        this.name = name;
-    }
-}
