+++ src/main/java/com/woncheonpeople/shoppingmall/domain/member/Member.java
@@ -1,19 +1,17 @@
 package com.woncheonpeople.shoppingmall.domain.member;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
-@Data
 @Entity
-@NoArgsConstructor
-@AllArgsConstructor
+@RequiredArgsConstructor
+@Getter @Setter
 public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column
     private String name;
 
     @Enumerated(EnumType.STRING)
@@ -22,9 +20,12 @@ public class Member {
     @Embedded
     private Address address;
 
+    @Column
     private String phoneNumber;
 
+    @Column
     private String loginId;
 
+    @Column
     private String password;
 }
