+++ src/main/java/com/woncheonpeople/shoppingmall/domain/product/ProductImage.java
@@ -1,25 +1,22 @@
-package com.woncheonpeople.shoppingmall.domain;
+package com.woncheonpeople.shoppingmall.domain.product;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
-@Entity(name = "order_items")
+@Entity
 @RequiredArgsConstructor
 @Getter @Setter
-public class OrderItem {
+public class ProductImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @ManyToOne(targetEntity = Order.class, fetch = FetchType.LAZY)
-    @JoinColumn(name = "order_id", updatable = false)
-    private Order order;
 
     @ManyToOne(targetEntity = Product.class, fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id", updatable = false)
     private Product product;
 
-
+    @Column
+    private String imageUrl;
 }
