+++ src/main/java/com/woncheonpeople/shoppingmall/domain/product/ProductOption.java
@@ -5,10 +5,9 @@ import jakarta.persistence.*;
 import lombok.*;
 
 
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter @Setter
 @Entity
+@RequiredArgsConstructor
+@Getter @Setter
 public class ProductOption {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
