--- src/main/java/com/woncheonpeople/shoppingmall/domain/product/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.woncheonpeople.shoppingmall.domain.product;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-    List<Product> findByName(String name);
-
-}
