--- src/main/java/com/woncheonpeople/shoppingmall/jparepository/MemberJpaRepository.java
@@ -1,7 +1,0 @@
-package com.woncheonpeople.shoppingmall.jparepository;
-
-import com.woncheonpeople.shoppingmall.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberJpaRepository extends JpaRepository<Member, Long> {
-}
