+++ src/main/java/com/woncheonpeople/shoppingmall/repository/OrderJpaRepository.java
@@ -1,6 +1,6 @@
-package com.woncheonpeople.shoppingmall.jparepository;
+package com.woncheonpeople.shoppingmall.repository;
 
-import com.woncheonpeople.shoppingmall.domain.Order;
+import com.woncheonpeople.shoppingmall.domain.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderJpaRepository extends JpaRepository<Order, Long> {
