+++ src/test/java/com/woncheonpeople/shoppingmall/repository/OrderJpaRepositoryTest.java
@@ -1,12 +1,8 @@
-package com.woncheonpeople.shoppingmall.jparepository;
+package com.woncheonpeople.shoppingmall.repository;
 
-import com.woncheonpeople.shoppingmall.domain.Member;
-import com.woncheonpeople.shoppingmall.domain.Product;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 public class OrderJpaRepositoryTest {
